fix: solve requested changes
