+++ object/permission.go
@@ -45,7 +45,7 @@ type Permission struct {
 }
 
 type PermissionRule struct {
-	PType string `xorm:"varchar(100) index not null default ''" json:"pType"`
+	Ptype string `xorm:"varchar(100) index not null default ''" json:"ptype"`
 	V0    string `xorm:"varchar(100) index not null default ''" json:"v0"`
 	V1    string `xorm:"varchar(100) index not null default ''" json:"v1"`
 	V2    string `xorm:"varchar(100) index not null default ''" json:"v2"`
